requirements.txt
google-generativeai
pypdf
streamlit